UCP/UCT: alignment parameter - apply additional CR comments.

+++ src/uct/api/uct.h
@@ -269,7 +269,7 @@ struct uct_iface_attr {
             size_t           max_zcopy;  /**< Maximal size for put_zcopy (total
                                               of @ref uct_iov_t::length of the
                                               @a iov parameter) */
-            uint16_t         opt_zcopy_align; /**< Alignment parameter */
+            size_t           opt_zcopy_align; /**< Alignment parameter */
             size_t           align_mtu;       /**< MTU used for alignment */
             size_t           max_iov;    /**< Maximal @a iovcnt parameter in
                                               @ref ::uct_ep_put_zcopy
@@ -284,7 +284,7 @@ struct uct_iface_attr {
             size_t           max_zcopy;  /**< Maximal size for get_zcopy (total
                                               of @ref uct_iov_t::length of the
                                               @a iov parameter) */
-            uint16_t         opt_zcopy_align; /**< Alignment parameter */
+            size_t           opt_zcopy_align; /**< Alignment parameter */
             size_t           align_mtu;       /**< MTU used for alignment */
             size_t           max_iov;    /**< Maximal @a iovcnt parameter in
                                               @ref uct_ep_get_zcopy
@@ -300,7 +300,7 @@ struct uct_iface_attr {
             size_t           max_zcopy;  /**< Total max. size (incl. the header
                                               and total of @ref uct_iov_t::length
                                               of the @a iov parameter) */
-            uint16_t         opt_zcopy_align; /**< Alignment parameter */
+            size_t           opt_zcopy_align; /**< Alignment parameter */
             size_t           align_mtu;       /**< MTU used for alignment */
             size_t           max_hdr;    /**< Max. header size for zcopy */
             size_t           max_iov;    /**< Maximal @a iovcnt parameter in
